api/settings: Move "PUT /settings/club" into "clubs" module
