+++ skylines/api/views/clubs.py
@@ -6,6 +6,10 @@ from skylines.database import db
 from skylines.api.oauth import oauth
 from skylines.lib.dbutil import get_requested_record
 from skylines.model import Club, User
+from skylines.model.event import (
+    create_club_join_event
+)
+
 from skylines.schemas import ClubSchema, ValidationError
 
 clubs_blueprint = Blueprint('clubs', 'skylines')
@@ -33,6 +37,42 @@ def read(club_id):
     json['isWritable'] = club.is_writable(current_user)
 
     return jsonify(json)
+
+
+@clubs_blueprint.route('/settings/club', methods=['PUT'])
+@oauth.required()
+def create_club():
+    current_user = User.get(request.user_id)
+    if not current_user:
+        return jsonify(error='invalid-token'), 401
+
+    json = request.get_json()
+    if json is None:
+        return jsonify(error='invalid-request'), 400
+
+    try:
+        data = ClubSchema(only=('name',)).load(json).data
+    except ValidationError, e:
+        return jsonify(error='validation-failed', fields=e.messages), 422
+
+    if Club.exists(name=data.get('name')):
+        return jsonify(error='duplicate-club-name'), 422
+
+    # create the new club
+    club = Club(**data)
+    club.owner_id = current_user.id
+    db.session.add(club)
+    db.session.flush()
+
+    # assign the user to the new club
+    current_user.club = club
+
+    # create the "user joined club" event
+    create_club_join_event(club.id, current_user)
+
+    db.session.commit()
+
+    return jsonify(id=club.id)
 
 
 @clubs_blueprint.route('/clubs/<club_id>', methods=['POST'], strict_slashes=False)
